+++ src/main/kotlin/net/dstribe/customize_word_separators/PrevWordWithSelectionAction.kt
@@ -1,0 +1,19 @@
+package net.dstribe.customize_word_separators
+
+import com.intellij.openapi.actionSystem.AnAction
+import com.intellij.openapi.actionSystem.AnActionEvent
+import com.intellij.openapi.actionSystem.CommonDataKeys
+import com.intellij.openapi.editor.Editor
+import com.intellij.openapi.editor.actionSystem.EditorActionHandler
+
+class PrevWordWithSelectionAction : AnAction() {
+    override fun actionPerformed(e: AnActionEvent) {
+        val editor: Editor = e.getRequiredData(CommonDataKeys.EDITOR)
+        val actionHandler: EditorActionHandler = NextPrevWordHandler(
+            myNext = false,
+            myWithSelection = true,
+            e = e
+        )
+        actionHandler.execute(editor, null, e.dataContext)
+    }
+}